requirements.txt
attrs==17.4.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coverage==4.0.3
dj-database-url==0.5.0
Django==2.0.2
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
idna-ssl==1.0.0
multidict==4.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycparser==2.18
PyJWT==1.6.0
python-coveralls==2.9.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
django-cors-headers
